Refactor services into consolidated modules
