--- SaashrConfig.cs
@@ -1,7 +1,0 @@
-namespace ClockEnforcer
-{
-    internal static class SaashrConfig
-    {
-        public const string ApiKey = "iibh7b86dlces64stxqwm15n65kvkhf3";
-    }
-}
